refactor runtime
