+++ azure-toolkit-libs/azure-toolkit-appservice-lib/src/main/java/com/microsoft/azure/toolkit/lib/appservice/model/FunctionAppRuntime.java
@@ -1,9 +1,7 @@
 package com.microsoft.azure.toolkit.lib.appservice.model;
 
-import com.azure.resourcemanager.appservice.models.JavaVersion;
 import org.apache.commons.lang3.StringUtils;
 
-import javax.annotation.Nonnull;
 import java.util.List;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -13,27 +11,11 @@ public interface FunctionAppRuntime extends Runtime {
     FunctionAppRuntime DOCKER = FunctionAppDockerRuntime.INSTANCE;
     FunctionAppRuntime DEFAULT = FunctionAppWindowsRuntime.FUNCTION_JAVA17;
 
-    /**
-     * @return java version number, e.g. '1.8'(windows only), '8', '11', '17'
-     */
-    @Nonnull
-    String getJavaVersionNumber();
-
     default String getDisplayName() {
         if (this.isDocker()) {
             return "Docker";
         }
         return String.format("%s-%s", this.getOperatingSystem().toString(), this.getJavaVersionUserText());
-    }
-
-    default String getJavaVersionUserText() {
-        if (this.isDocker()) {
-            return "";
-        }
-        if (StringUtils.equalsAnyIgnoreCase(getJavaVersionNumber(), "1.8", "8")) {
-            return "Java 8";
-        }
-        return String.format("Java %s", getJavaVersionNumber());
     }
 
     @Override
@@ -43,14 +25,6 @@ public interface FunctionAppRuntime extends Runtime {
         }
         final Pattern EXCLUDE_PATTERN = Pattern.compile("\\..*\\.");
         return EXCLUDE_PATTERN.matcher(this.getJavaVersionNumber()).matches();
-    }
-
-    @Override
-    default JavaVersion getJavaVersion() {
-        if (this.isDocker()) {
-            return JavaVersion.fromString("docker");
-        }
-        return JavaVersion.fromString(this.getJavaVersionNumber());
     }
 
     static List<FunctionAppRuntime> getMajorRuntimes() {
